Refactoring to microservice architecture, phase 5 (Inventory-service): - inventory-service for Books refactored, jooq -> JPA - inventory-service for ShoppingCart refactored, jooq -> JPA - inventory-service for Orders refactored, jooq -> JPA

+++ common/src/main/kotlin/dto/PaginatedDataResponse.kt
@@ -1,10 +1,9 @@
 package dto
 
-
 data class PaginatedDataResponse<T>(
     val data: List<T>,
     val totalElements: Long,
-    val totalPages: Long,
+    val totalPages: Int,
     val currentPage: Int,
     val pageSize: Int
 )
